Validate observation is now available on all observation list views
