+++ frontend/src/app/auth/user-dashboard/user-observations/user-observations.component.ts
@@ -31,6 +31,7 @@ export class UserObsComponent extends ObsComponent implements OnInit {
     userDashboard = true;
 
     ngOnInit() {
+        this.userObservations = this.myObs
         this.breakpointObserver
             .observe(['(min-width: 700px)'])
             .subscribe((state: BreakpointState) => {
